refactor(dashboard): move ViewSwitcher into its own component file
Relocate the standalone ViewSwitcher component (and its ViewMode type) out of
App.tsx into components/ViewSwitcher.tsx — a pure file move, no JSX or logic
change. App imports it; the SplitDirection import (now unused in App) is
pruned.

Part of #5560 (App.tsx Shell decomposition).
